Due to bugs fix built snapshot versions are used for some plugins.    The cargo-maven2-plugin has further a patch applied.
Repositories reorganized. Plugin repos are the same as the normal repos.
Release repos:
   http://el4.elca-services.ch/el4j/maven2repository
   http://leaffy.elca.ch/java/maven2repository
Snapshot repos:
   http://el4.elca-services.ch/el4j/maven2snapshots
   http://leaffy.elca.ch/java/maven2snapshots

EL4J plugin and module version are now all marked as SNAPSHOT and no more as "alpha-1". This is made to easily deploy a snapshot on the correct repository and prevent the user to deploy a snapshot in the release repo.