Remove the getter for ConvenienceHibernateTemplate from the interface.
Here the snipped of the mail, why we removed the method:
About the ConvenienceHibernateTemplate: It really makes no sense to have this method on the interface, but the getter and setter must stay public because the base property "hibernateTemplate" of Spring class "org.springframework.orm.hibernate3.support.HibernateDaoSupport" is also public. When we make now our property "ConvenienceHibernateTemplate" protected the bad developers will start using property "hibernateTemplate".
But by removing the method from the interface I think we already have achieved the purpose.